+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 1.0.2
-*
+* Add optional build number argument to console command #44
 
 ### 1.0.1
 * Add crumb request header because new Jenkins >= 2.x installation enables CSRF protection by default #40